--- arr.js
@@ -1,662 +1,0 @@
-const arrKey = [
-  {
-    'en': [
-      [
-        {
-          'small': '`',
-          'shift': '~',
-          'code': 'Backquote',
-        },
-        {
-          'small': '1',
-          'shift': '!',
-          'code': 'Digit1',
-        },
-        {
-          'small': '2',
-          'shift': '@',
-          'code': 'Digit2',
-        },
-        {
-          'small': '3',
-          'shift': '#',
-          'code': 'Digit3',
-        },
-        {
-          'small': '4',
-          'shift': '$',
-          'code': 'Digit4',
-        },
-        {
-          'small': '5',
-          'shift': '%',
-          'code': 'Digit5',
-        },
-        {
-          'small': '6',
-          'shift': '^',
-          'code': 'Digit6',
-        },
-        {
-          'small': '7',
-          'shift': '&',
-          'code': 'Digit7',
-        },
-        {
-          'small': '8',
-          'shift': '*',
-          'code': 'Digit8',
-        },
-        {
-          'small': '9',
-          'shift': '(',
-          'code': 'Digit9',
-        },
-        {
-          'small': '0',
-          'shift': ')',
-          'code': 'Digit0',
-        },
-        {
-          'small': '-',
-          'shift': '_',
-          'code': 'Minus',
-        },
-        {
-          'small': '=',
-          'shift': '+',
-          'code': 'Equal',
-        },
-        {
-          'small': 'Backspace',
-          'shift': null,
-          'code': 'Backspace',
-        },
-      ],
-      [
-        {
-          'small': 'Tab',
-          'shift': null,
-          'code': 'Tab',
-        },
-        {
-          'small': 'q',
-          'shift': 'Q',
-          'code': 'KeyQ',
-        },
-        {
-          'small': 'w',
-          'shift': 'W',
-          'code': 'KeyW',
-        },
-        {
-          'small': 'e',
-          'shift': 'E',
-          'code': 'KeyE',
-        },
-        {
-          'small': 'r',
-          'shift': 'R',
-          'code': 'KeyR',
-        },
-        {
-          'small': 't',
-          'shift': 'T',
-          'code': 'KeyT',
-        },
-        {
-          'small': 'y',
-          'shift': 'Y',
-          'code': 'KeyY',
-        },
-        {
-          'small': 'u',
-          'shift': 'U',
-          'code': 'KeyU',
-        },
-        {
-          'small': 'i',
-          'shift': 'I',
-          'code': 'KeyI',
-        },
-        {
-          'small': 'o',
-          'shift': 'O',
-          'code': 'KeyO',
-        },
-        {
-          'small': 'p',
-          'shift': 'P',
-          'code': 'KeyP',
-        },
-        {
-          'small': '[',
-          'shift': '{',
-          'code': 'BracketLeft',
-        },
-        {
-          'small': ']',
-          'shift': '}',
-          'code': 'BracketRight',
-        },
-        {
-          'small': '\\',
-          'shift': '|',
-          'code': 'Backslash',
-        },
-        {
-          'small': 'Delete',
-          'shift': null,
-          'code': 'Delete',
-        },
-      ],
-      [
-        {
-          'small': 'CapsLock',
-          'shift': null,
-          'code': 'CapsLock',
-        },
-        {
-          'small': 'a',
-          'shift': 'A',
-          'code': 'KeyA',
-        },
-        {
-          'small': 's',
-          'shift': 'S',
-          'code': 'KeyS',
-        },
-        {
-          'small': 'd',
-          'shift': 'D',
-          'code': 'KeyD',
-        },
-        {
-          'small': 'f',
-          'shift': 'F',
-          'code': 'KeyF',
-        },
-        {
-          'small': 'g',
-          'shift': 'G',
-          'code': 'KeyG',
-        },
-        {
-          'small': 'h',
-          'shift': 'H',
-          'code': 'KeyH',
-        },
-        {
-          'small': 'j',
-          'shift': 'J',
-          'code': 'KeyJ',
-        },
-        {
-          'small': 'k',
-          'shift': 'K',
-          'code': 'KeyK',
-        },
-        {
-          'small': 'l',
-          'shift': 'L',
-          'code': 'KeyL',
-        },
-        {
-          'small': ';',
-          'shift': ':',
-          'code': 'Semicolon',
-        },
-        {
-          'small': "'",
-          'shift': '"',
-          'code': 'Quote',
-        },
-        {
-          'small': 'Enter',
-          'shift': null,
-          'code': 'Enter',
-        },
-      ],
-      [
-        {
-          'small': 'Shift',
-          'shift': null,
-          'code': 'ShiftLeft',
-        },
-        {
-          'small': 'z',
-          'shift': 'Z',
-          'code': 'KeyZ',
-        },
-        {
-          'small': 'x',
-          'shift': 'X',
-          'code': 'KeyX',
-        },
-        {
-          'small': 'c',
-          'shift': 'C',
-          'code': 'KeyC',
-        },
-        {
-          'small': 'v',
-          'shift': 'V',
-          'code': 'KeyV',
-        },
-        {
-          'small': 'b',
-          'shift': 'B',
-          'code': 'KeyB',
-        },
-        {
-          'small': 'n',
-          'shift': 'N',
-          'code': 'KeyN',
-        },
-        {
-          'small': 'm',
-          'shift': 'M',
-          'code': 'KeyM',
-        },
-        {
-          'small': ',',
-          'shift': '<',
-          'code': 'Comma',
-        },
-        {
-          'small': '.',
-          'shift': '>',
-          'code': 'Period',
-        },
-        {
-          'small': '/',
-          'shift': '?',
-          'code': 'Slash',
-        },
-        {
-          'small': '&uarr;',
-          'shift': null,
-          'code': 'ArrowUp',
-        },
-        {
-          'small': 'Shift',
-          'shift': null,
-          'code': 'ShiftRight',
-        },
-      ],
-      [
-        {
-          'small': 'Ctrl',
-          'shift': null,
-          'code': 'ControlLeft',
-        },
-        {
-          'small': 'Alt',
-          'shift': null,
-          'code': 'AltLeft',
-        },
-        {
-          'small': ' ',
-          'shift': null,
-          'code': 'Space',
-        },
-        {
-          'small': 'Alt',
-          'shift': null,
-          'code': 'AltRight',
-        },
-        {
-          'small': '&larr;',
-          'shift': null,
-          'code': 'ArrowLeft',
-        },
-        {
-          'small': '&darr;',
-          'shift': null,
-          'code': 'ArrowDown',
-        },
-        {
-          'small': '&rarr;',
-          'shift': null,
-          'code': 'ArrowRight',
-        },
-        {
-          'small': 'Ctrl',
-          'shift': null,
-          'code': 'ControlRight',
-        },
-      ],
-    ],
-  },
-  {
-    'ru': [
-      [
-        {
-          'small': 'ё',
-          'shift': 'Ё',
-          'code': 'Backquote',
-        },
-        {
-          'small': '1',
-          'shift': '!',
-          'code': 'Digit1',
-        },
-        {
-          'small': '2',
-          'shift': '"',
-          'code': 'Digit2',
-        },
-        {
-          'small': '3',
-          'shift': '№',
-          'code': 'Digit3',
-        },
-        {
-          'small': '4',
-          'shift': ';',
-          'code': 'Digit4',
-        },
-        {
-          'small': '5',
-          'shift': '%',
-          'code': 'Digit5',
-        },
-        {
-          'small': '6',
-          'shift': ':',
-          'code': 'Digit6',
-        },
-        {
-          'small': '7',
-          'shift': '?',
-          'code': 'Digit7',
-        },
-        {
-          'small': '8',
-          'shift': '*',
-          'code': 'Digit8',
-        },
-        {
-          'small': '9',
-          'shift': '(',
-          'code': 'Digit9',
-        },
-        {
-          'small': '0',
-          'shift': ')',
-          'code': 'Digit0',
-        },
-        {
-          'small': '-',
-          'shift': '_',
-          'code': 'Minus',
-        },
-        {
-          'small': '=',
-          'shift': '+',
-          'code': 'Equal',
-        },
-        {
-          'small': 'Backspace',
-          'shift': null,
-          'code': 'Backspace',
-        },
-      ],
-      [
-        {
-          'small': 'Tab',
-          'shift': null,
-          'code': 'Tab',
-        },
-        {
-          'small': 'й',
-          'shift': 'Й',
-          'code': 'KeyQ',
-        },
-        {
-          'small': 'ц',
-          'shift': 'Ц',
-          'code': 'KeyW',
-        },
-        {
-          'small': 'у',
-          'shift': 'У',
-          'code': 'KeyE',
-        },
-        {
-          'small': 'к',
-          'shift': 'К',
-          'code': 'KeyR',
-        },
-        {
-          'small': 'е',
-          'shift': 'Е',
-          'code': 'KeyT',
-        },
-        {
-          'small': 'н',
-          'shift': 'Н',
-          'code': 'KeyY',
-        },
-        {
-          'small': 'г',
-          'shift': 'Г',
-          'code': 'KeyU',
-        },
-        {
-          'small': 'ш',
-          'shift': 'Ш',
-          'code': 'KeyI',
-        },
-        {
-          'small': 'щ',
-          'shift': 'Щ',
-          'code': 'KeyO',
-        },
-        {
-          'small': 'з',
-          'shift': 'З',
-          'code': 'KeyP',
-        },
-        {
-          'small': 'х',
-          'shift': 'Х',
-          'code': 'BracketLeft',
-        },
-        {
-          'small': 'ъ',
-          'shift': 'Ъ',
-          'code': 'BracketRight',
-        },
-        {
-          'small': '\\',
-          'shift': '/',
-          'code': 'Backslash',
-        },
-        {
-          'small': 'Delete',
-          'shift': null,
-          'code': 'Delete',
-        },
-      ],
-      [
-        {
-          'small': 'CapsLock',
-          'shift': null,
-          'code': 'CapsLock',
-        },
-        {
-          'small': 'ф',
-          'shift': 'Ф',
-          'code': 'KeyA',
-        },
-        {
-          'small': 'ы',
-          'shift': 'Ы',
-          'code': 'KeyS',
-        },
-        {
-          'small': 'в',
-          'shift': 'В',
-          'code': 'KeyD',
-        },
-        {
-          'small': 'а',
-          'shift': 'А',
-          'code': 'KeyF',
-        },
-        {
-          'small': 'п',
-          'shift': 'П',
-          'code': 'KeyG',
-        },
-        {
-          'small': 'р',
-          'shift': 'Р',
-          'code': 'KeyH',
-        },
-        {
-          'small': 'о',
-          'shift': 'О',
-          'code': 'KeyJ',
-        },
-        {
-          'small': 'л',
-          'shift': 'Л',
-          'code': 'KeyK',
-        },
-        {
-          'small': 'д',
-          'shift': 'Д',
-          'code': 'KeyL',
-        },
-        {
-          'small': 'ж',
-          'shift': 'Ж',
-          'code': 'Semicolon',
-        },
-        {
-          'small': 'э',
-          'shift': 'Э',
-          'code': 'Quote',
-        },
-        {
-          'small': 'Enter',
-          'shift': null,
-          'code': 'Enter',
-        },
-      ],
-      [
-        {
-          'small': 'Shift',
-          'shift': null,
-          'code': 'ShiftLeft',
-        },
-        {
-          'small': 'я',
-          'shift': 'Я',
-          'code': 'KeyZ',
-        },
-        {
-          'small': 'ч',
-          'shift': 'Ч',
-          'code': 'KeyX',
-        },
-        {
-          'small': 'c',
-          'shift': 'C',
-          'code': 'KeyC',
-        },
-        {
-          'small': 'м',
-          'shift': 'М',
-          'code': 'KeyV',
-        },
-        {
-          'small': 'и',
-          'shift': 'И',
-          'code': 'KeyB',
-        },
-        {
-          'small': 'т',
-          'shift': 'Т',
-          'code': 'KeyN',
-        },
-        {
-          'small': 'ь',
-          'shift': 'Ь',
-          'code': 'KeyM',
-        },
-        {
-          'small': 'б',
-          'shift': 'Б',
-          'code': 'Comma',
-        },
-        {
-          'small': 'ю',
-          'shift': 'Ю',
-          'code': 'Period',
-        },
-        {
-          'small': '.',
-          'shift': ',',
-          'code': 'Slash',
-        },
-        {
-          'small': '&uarr;',
-          'shift': null,
-          'code': 'ArrowUp',
-        },
-        {
-          'small': 'Shift',
-          'shift': null,
-          'code': 'ShiftRight',
-        },
-      ],
-      [
-        {
-          'small': 'Ctrl',
-          'shift': null,
-          'code': 'ControlLeft',
-        },
-        {
-          'small': 'Alt',
-          'shift': null,
-          'code': 'AltLeft',
-        },
-        {
-          'small': ' ',
-          'shift': null,
-          'code': 'Space',
-        },
-        {
-          'small': 'Alt',
-          'shift': null,
-          'code': 'AltRight',
-        },
-        {
-          'small': '&larr;',
-          'shift': null,
-          'code': 'ArrowLeft',
-        },
-        {
-          'small': '&darr;',
-          'shift': null,
-          'code': 'ArrowDown',
-        },
-        {
-          'small': '&rarr;',
-          'shift': null,
-          'code': 'ArrowRight',
-        },
-        {
-          'small': 'Ctrl',
-          'shift': null,
-          'code': 'ControlRight',
-        },
-      ],
-    ],
-  },
-];
-
-export default arrKey;
