+++ index.html
@@ -16,6 +16,6 @@
     />
   </head>
   <body>
-    <script type="module" src="script.js"></script>
+    <script type="module" src="./scripts/script.js"></script>
   </body>
 </html>
