+++ scripts/arr.js
@@ -1,0 +1,718 @@
+const arrKey = [
+  [
+    {
+      'en': {
+        'small': '`',
+        'shift': '~',
+      },
+      'ru': {
+        'small': 'ё',
+        'shift': 'Ё',
+      },
+      'code': 'Backquote',
+    },
+    {
+      'en': {
+        'small': '1',
+        'shift': '!',
+      },
+      'ru': {
+        'small': '1',
+        'shift': '!',
+      },
+      'code': 'digit1',
+    },
+    {
+      'en': {
+        'small': '2',
+        'shift': '@',
+      },
+      'ru': {
+        'small': '2',
+        'shift': '"',
+      },
+      'code': 'digit2',
+    },
+    {
+      'en': {
+        'small': '3',
+        'shift': '#',
+      },
+      'ru': {
+        'small': '3',
+        'shift': '№',
+      },
+      'code': 'digit3',
+    },
+    {
+      'en': {
+        'small': '4',
+        'shift': '$',
+      },
+      'ru': {
+        'small': '4',
+        'shift': ';',
+      },
+      'code': 'digit4',
+    },
+    {
+      'en': {
+        'small': '5',
+        'shift': '%',
+      },
+      'ru': {
+        'small': '5',
+        'shift': '%',
+      },
+      'code': 'digit5',
+    },
+    {
+      'en': {
+        'small': '6',
+        'shift': '^',
+      },
+      'ru': {
+        'small': '6',
+        'shift': ':',
+      },
+      'code': 'digit6',
+    },
+    {
+      'en': {
+        'small': '7',
+        'shift': '&',
+      },
+      'ru': {
+        'small': '7',
+        'shift': '?',
+      },
+      'code': 'digit7',
+    },
+    {
+      'en': {
+        'small': '8',
+        'shift': '*',
+      },
+      'ru': {
+        'small': '8',
+        'shift': '*',
+      },
+      'code': 'digit8',
+    },
+    {
+      'en': {
+        'small': '9',
+        'shift': '(',
+      },
+      'ru': {
+        'small': '9',
+        'shift': '(',
+      },
+      'code': 'digit9',
+    },
+    {
+      'en': {
+        'small': '0',
+        'shift': ')',
+      },
+      'ru': {
+        'small': '0',
+        'shift': ')',
+      },
+      'code': 'digit0',
+    },
+    {
+      'en': {
+        'small': '-',
+        'shift': '_',
+      },
+      'ru': {
+        'small': '-',
+        'shift': '_',
+      },
+      'code': 'Minus',
+    },
+    {
+      'en': {
+        'small': '=',
+        'shift': '+',
+      },
+      'ru': {
+        'small': '=',
+        'shift': '+',
+      },
+      'code': 'Equal',
+    },
+    {
+      'en': {
+        'small': 'Backspace',
+        'shift': 'Backspace',
+      },
+      'ru': {
+        'small': 'Backspace',
+        'shift': 'Backspace',
+      },
+      'code': 'Backspace',
+    },
+  ],
+  [
+    {
+      'en': {
+        'small': 'Tab',
+        'shift': 'Tab',
+      },
+      'ru': {
+        'small': 'Tab',
+        'shift': 'Tab',
+      },
+      'code': 'Tab',
+    },
+    {
+      'en': {
+        'small': 'q',
+        'shift': 'Q',
+      },
+      'ru': {
+        'small': 'й',
+        'shift': 'Й',
+      },
+      'code': 'KeyQ',
+    },
+    {
+      'en': {
+        'small': 'w',
+        'shift': 'W',
+      },
+      'ru': {
+        'small': 'ц',
+        'shift': 'Ц',
+      },
+      'code': 'KeyW',
+    },
+    {
+      'en': {
+        'small': 'e',
+        'shift': 'E',
+      },
+      'ru': {
+        'small': 'у',
+        'shift': 'У',
+      },
+      'code': 'KeyE',
+    },
+    {
+      'en': {
+        'small': 'r',
+        'shift': 'R',
+      },
+      'ru': {
+        'small': 'к',
+        'shift': 'К',
+      },
+      'code': 'KeyR',
+    },
+    {
+      'en': {
+        'small': 't',
+        'shift': 'T',
+      },
+      'ru': {
+        'small': 'е',
+        'shift': 'Е',
+      },
+      'code': 'KeyT',
+    },
+    {
+      'en': {
+        'small': 'y',
+        'shift': 'Y',
+      },
+      'ru': {
+        'small': 'н',
+        'shift': 'Н',
+      },
+      'code': 'KeyY',
+    },
+    {
+      'en': {
+        'small': 'u',
+        'shift': 'U',
+      },
+      'ru': {
+        'small': 'г',
+        'shift': 'Г',
+      },
+      'code': 'KeyU',
+    },
+    {
+      'en': {
+        'small': 'i',
+        'shift': 'I',
+      },
+      'ru': {
+        'small': 'ш',
+        'shift': 'Ш',
+      },
+      'code': 'KeyI',
+    },
+    {
+      'en': {
+        'small': 'o',
+        'shift': 'O',
+      },
+      'ru': {
+        'small': 'щ',
+        'shift': 'Щ',
+      },
+      'code': 'KeyO',
+    },
+    {
+      'en': {
+        'small': 'p',
+        'shift': 'P',
+      },
+      'ru': {
+        'small': 'з',
+        'shift': 'З',
+      },
+      'code': 'KeyP',
+    },
+    {
+      'en': {
+        'small': '[',
+        'shift': '{',
+      },
+      'ru': {
+        'small': 'х',
+        'shift': 'Х',
+      },
+      'code': 'BracketLeft',
+    },
+    {
+      'en': {
+        'small': ']',
+        'shift': '}',
+      },
+      'ru': {
+        'small': 'ъ',
+        'shift': 'Ъ',
+      },
+      'code': 'BracketRight',
+    },
+    {
+      'en': {
+        'small': '\\',
+        'shift': '|',
+      },
+      'ru': {
+        'small': '\\',
+        'shift': '/',
+      },
+      'code': 'Backslash',
+    },
+    {
+      'en': {
+        'small': 'Delete',
+        'shift': 'Delete',
+      },
+      'ru': {
+        'small': 'Delete',
+        'shift': 'Delete',
+      },
+      'code': 'Delete',
+    },
+  ],
+  [
+    {
+      'en': {
+        'small': 'CapsLock',
+        'shift': 'CapsLock',
+      },
+      'ru': {
+        'small': 'CapsLock',
+        'shift': 'CapsLock',
+      },
+      'code': 'CapsLock',
+    },
+    {
+      'en': {
+        'small': 'a',
+        'shift': 'A',
+      },
+      'ru': {
+        'small': 'ф',
+        'shift': 'Ф',
+      },
+      'code': 'KeyA',
+    },
+    {
+      'en': {
+        'small': 's',
+        'shift': 'S',
+      },
+      'ru': {
+        'small': 'ы',
+        'shift': 'Ы',
+      },
+      'code': 'KeyS',
+    },
+    {
+      'en': {
+        'small': 'd',
+        'shift': 'D',
+      },
+      'ru': {
+        'small': 'в',
+        'shift': 'В',
+      },
+      'code': 'KeyD',
+    },
+    {
+      'en': {
+        'small': 'f',
+        'shift': 'F',
+      },
+      'ru': {
+        'small': 'а',
+        'shift': 'А',
+      },
+      'code': 'KeyF',
+    },
+    {
+      'en': {
+        'small': 'g',
+        'shift': 'G',
+      },
+      'ru': {
+        'small': 'п',
+        'shift': 'П',
+      },
+      'code': 'KeyG',
+    },
+    {
+      'en': {
+        'small': 'h',
+        'shift': 'H',
+      },
+      'ru': {
+        'small': 'р',
+        'shift': 'Р',
+      },
+      'code': 'KeyH',
+    },
+    {
+      'en': {
+        'small': 'j',
+        'shift': 'J',
+      },
+      'ru': {
+        'small': 'о',
+        'shift': 'О',
+      },
+      'code': 'KeyJ',
+    },
+    {
+      'en': {
+        'small': 'k',
+        'shift': 'K',
+      },
+      'ru': {
+        'small': 'л',
+        'shift': 'Л',
+      },
+      'code': 'KeyK',
+    },
+    {
+      'en': {
+        'small': 'l',
+        'shift': 'L',
+      },
+      'ru': {
+        'small': 'д',
+        'shift': 'Д',
+      },
+      'code': 'KeyL',
+    },
+    {
+      'en': {
+        'small': ';',
+        'shift': ':',
+      },
+      'ru': {
+        'small': 'ж',
+        'shift': 'Ж',
+      },
+      'code': 'Semicolon',
+    },
+    {
+      'en': {
+        'small': "'",
+        'shift': '"',
+      },
+      'ru': {
+        'small': 'э',
+        'shift': 'Э',
+      },
+      'code': 'Quote',
+    },
+    {
+      'en': {
+        'small': 'Enter',
+        'shift': 'Enter',
+      },
+      'ru': {
+        'small': 'Enter',
+        'shift': 'Enter',
+      },
+      'code': 'Enter',
+    },
+  ],
+  [
+    {
+      'en': {
+        'small': 'Shift',
+        'shift': 'Shift',
+      },
+      'ru': {
+        'small': 'Shift',
+        'shift': 'Shift',
+      },
+      'code': 'ShiftLeft',
+    },
+    {
+      'en': {
+        'small': 'z',
+        'shift': 'Z',
+      },
+      'ru': {
+        'small': 'я',
+        'shift': 'Я',
+      },
+      'code': 'KeyZ',
+    },
+    {
+      'en': {
+        'small': 'x',
+        'shift': 'X',
+      },
+      'ru': {
+        'small': 'ч',
+        'shift': 'Ч',
+      },
+      'code': 'KeyX',
+    },
+    {
+      'en': {
+        'small': 'c',
+        'shift': 'C',
+      },
+      'ru': {
+        'small': 'с',
+        'shift': 'С',
+      },
+      'code': 'KeyC',
+    },
+    {
+      'en': {
+        'small': 'v',
+        'shift': 'V',
+      },
+      'ru': {
+        'small': 'м',
+        'shift': 'М',
+      },
+      'code': 'KeyV',
+    },
+    {
+      'en': {
+        'small': 'b',
+        'shift': 'B',
+      },
+      'ru': {
+        'small': 'и',
+        'shift': 'И',
+      },
+      'code': 'KeyB',
+    },
+    {
+      'en': {
+        'small': 'n',
+        'shift': 'N',
+      },
+      'ru': {
+        'small': 'т',
+        'shift': 'Т',
+      },
+      'code': 'KeyN',
+    },
+    {
+      'en': {
+        'small': 'm',
+        'shift': 'M',
+      },
+      'ru': {
+        'small': 'ь',
+        'shift': 'Ь',
+      },
+      'code': 'KeyM',
+    },
+    {
+      'en': {
+        'small': ',',
+        'shift': '<',
+      },
+      'ru': {
+        'small': 'б',
+        'shift': 'Б',
+      },
+      'code': 'Comma',
+    },
+    {
+      'en': {
+        'small': '.',
+        'shift': '>',
+      },
+      'ru': {
+        'small': 'ю',
+        'shift': 'Ю',
+      },
+      'code': 'Period',
+    },
+    {
+      'en': {
+        'small': '/',
+        'shift': '?',
+      },
+      'ru': {
+        'small': '.',
+        'shift': ',',
+      },
+      'code': 'Slash',
+    },
+    {
+      'en': {
+        'small': '&#9650;',
+        'shift': '&#9650;',
+      },
+      'ru': {
+        'small': '&#9650;',
+        'shift': '&#9650;',
+      },
+      'code': 'ArrowUp',
+    },
+    {
+      'en': {
+        'small': 'Shift',
+        'shift': 'Shift',
+      },
+      'ru': {
+        'small': 'Shift',
+        'shift': 'Shift',
+      },
+      'code': 'ShiftRight',
+    },
+  ],
+  [
+    {
+      'en': {
+        'small': 'Ctrl',
+        'shift': 'Ctrl',
+      },
+      'ru': {
+        'small': 'Ctrl',
+        'shift': 'Ctrl',
+      },
+      'code': 'ControlLeft',
+    },
+    {
+      'en': {
+        'small': 'Win',
+        'shift': 'Win',
+      },
+      'ru': {
+        'small': 'Win',
+        'shift': 'Win',
+      },
+      'code': 'MetaLeft',
+    },
+    {
+      'en': {
+        'small': 'Alt',
+        'shift': 'Alt',
+      },
+      'ru': {
+        'small': 'Alt',
+        'shift': 'Alt',
+      },
+      'code': 'AltLeft',
+    },
+    {
+      'en': {
+        'small': '&ensp;',
+        'shift': '&ensp;',
+      },
+      'ru': {
+        'small': '&ensp;',
+        'shift': '&ensp;',
+      },
+      'code': 'Space',
+    },
+    {
+      'en': {
+        'small': 'Alt',
+        'shift': 'Alt',
+      },
+      'ru': {
+        'small': 'Alt',
+        'shift': 'Alt',
+      },
+      'code': 'AltRight',
+    },
+    {
+      'en': {
+        'small': '&#9668;',
+        'shift': '&#9668;',
+      },
+      'ru': {
+        'small': '&#9668;',
+        'shift': '&#9668;',
+      },
+      'code': 'ArrowLeft',
+    },
+    {
+      'en': {
+        'small': '&#9660;',
+        'shift': '&#9660;',
+      },
+      'ru': {
+        'small': '&#9660;',
+        'shift': '&#9660;',
+      },
+      'code': 'ArrowDown',
+    },
+    {
+      'en': {
+        'small': '&#9658;',
+        'shift': '&#9658;',
+      },
+      'ru': {
+        'small': '&#9658;',
+        'shift': '&#9658;',
+      },
+      'code': 'ArrowRight',
+    },
+    {
+      'en': {
+        'small': 'Ctrl',
+        'shift': 'Ctrl',
+      },
+      'ru': {
+        'small': 'Ctrl',
+        'shift': 'Ctrl',
+      },
+      'code': 'ControlRight',
+    },
+  ],
+];
+
+export default arrKey;
