+++ style.css
@@ -3,6 +3,9 @@
   --text-modal: #000;
   --bg-color: rgba(0, 0, 0, 0.25);
   --main-color: rgb(45, 52, 71);
+  --btn-color: rgb(69, 78, 105);
+  --keyBoard-color: rgb(180, 191, 203);
+  --textArea-color: rgb(236, 238, 224);
 }
 
 * {
@@ -32,20 +35,74 @@ body {
 
 .header h1 {
   display: block;
-  font-size: 2rem;
+  font-size: 1.8rem;
   text-align: center;
+  margin: 0.6rem 0;
 }
 
 .main {
   background-color: var(--bg-color);
   flex: 1 1 auto;
   display: flex;
+  flex-direction: column;
   align-items: center;
+}
+
+.main__container {
+  padding: 20px 0;
+  height: 90%;
+  width: 950px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.main__input {
+  background-color: var(--textArea-color);
+  border-radius: 3px;
+  box-shadow: 5px 5px 10px rgba(0, 0, 0, 0.5);
+}
+
+.main__keyboard {
+  width: 100%;
+  height: 320px;
+  margin: 0 auto;
+  padding: 10px 10px;
+  border-radius: 5px;
+  background-color: var(--keyBoard-color);
+  border: 3px solid var(--main-color);
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  box-shadow: 5px 5px 10px rgba(0, 0, 0, 0.5);
+}
+
+.key {
+  width: 50px;
+  height: 50px;
+  background-color: var(--btn-color);
+  font-weight: 700;
+  line-height: 50px;
+  text-align: center;
+  border-bottom: 2px solid var(--text-color);
+  border-right: 2px solid var(--text-color);
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.shift-key {
+  background-color: blue;
+}
+
+.row {
+  display: flex;
+  justify-content: space-between;
+  flex-wrap: nowrap;
 }
 
 .footer {
   background-color: var(--main-color);
-  padding: 1rem 0;
+  padding: 0.8rem 0;
   display: flex;
   flex-direction: column;
   align-items: center;
